chore(unity-react-core): add individual build files for each component
